start testing serverless run function 2
